+++ src/Symfony/Bridge/Doctrine/Tests/Security/RememberMe/DoctrineTokenProviderTest.php
@@ -56,6 +56,56 @@ class DoctrineTokenProviderTest extends TestCase
         $provider->loadTokenBySeries('someSeries');
     }
 
+    public function testVerifyOutdatedTokenAfterParallelRequest()
+    {
+        $provider = $this->bootstrapProvider();
+        $series = base64_encode(random_bytes(64));
+        $oldValue = 'oldValue';
+        $newValue = 'newValue';
+
+        // setup existing token
+        $token = new PersistentToken('someClass', 'someUser', $series, $oldValue, new \DateTime('2013-01-26T18:23:51'));
+        $provider->createNewToken($token);
+
+        // new request comes in requiring remember-me auth, which updates the token
+        $provider->updateExistingToken($token, $newValue, new \DateTime('-5 seconds'));
+        $provider->updateToken($series, $newValue, new \DateTime('-5 seconds'));
+
+        // parallel request comes in with the old remember-me cookie and session, which also requires reauth
+        $token = $provider->loadTokenBySeries($series);
+        $this->assertEquals($newValue, $token->getTokenValue());
+
+        // new token is valid
+        $this->assertTrue($provider->verifyToken($token, $newValue));
+        // old token is still valid
+        $this->assertTrue($provider->verifyToken($token, $oldValue));
+    }
+
+    public function testVerifyOutdatedTokenAfterParallelRequestFailsAfter60Seconds()
+    {
+        $provider = $this->bootstrapProvider();
+        $series = base64_encode(random_bytes(64));
+        $oldValue = 'oldValue';
+        $newValue = 'newValue';
+
+        // setup existing token
+        $token = new PersistentToken('someClass', 'someUser', $series, $oldValue, new \DateTime('2013-01-26T18:23:51'));
+        $provider->createNewToken($token);
+
+        // new request comes in requiring remember-me auth, which updates the token
+        $provider->updateExistingToken($token, $newValue, new \DateTime('-61 seconds'));
+        $provider->updateToken($series, $newValue, new \DateTime('-5 seconds'));
+
+        // parallel request comes in with the old remember-me cookie and session, which also requires reauth
+        $token = $provider->loadTokenBySeries($series);
+        $this->assertEquals($newValue, $token->getTokenValue());
+
+        // new token is valid
+        $this->assertTrue($provider->verifyToken($token, $newValue));
+        // old token is not valid anymore after 60 seconds
+        $this->assertFalse($provider->verifyToken($token, $oldValue));
+    }
+
     /**
      * @return DoctrineTokenProvider
      */